supporting-layer/uiux-authoring-tool/requirements.txt

django>=1.10.4 
djangorestframework
markdown
django-filter